requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
PyPDF2
streamlit_pdf_viewer
pytesseract
pdf2image
poppler-utils